Make race easier to start
Most players seem to have no idea which button is ability 1, so change to interact.

Rather than disallowing buttons to avoid glitches, stun the player so they are foced to let go of actions. Players seem to like to try to start while swinging around in gather room, so this was being confusing.

This has the benefit of not needing allow/disallow buttons at any point. This can be entirely controlled by regular custom game config instead of in workshop.